setup/requirements.txt
langchain-core
langchain-google-genai
langgraph
python-dotenv
spotipy
google-cloud-secret-manager
grpcio==1.67.1
streamlit